+++ src/environment/python_status.rs
@@ -1,0 +1,62 @@
+use std::path::Path;
+
+use rattler::install::{PythonInfo, Transaction};
+use rattler_conda_types::{PrefixRecord, RepoDataRecord};
+
+#[derive(Clone, Debug)]
+pub enum PythonStatus {
+    /// The python interpreter changed from `old` to `new`.
+    Changed { old: PythonInfo, new: PythonInfo },
+
+    /// The python interpreter remained the same.
+    Unchanged(PythonInfo),
+
+    /// The python interpreter was removed from the environment
+    Removed { old: PythonInfo },
+
+    /// The python interpreter was added to the environment
+    Added { new: PythonInfo },
+
+    /// There is no python interpreter in the environment.
+    DoesNotExist,
+}
+
+impl PythonStatus {
+    /// Determine the [`PythonStatus`] from a [`Transaction`].
+    pub(crate) fn from_transaction(
+        transaction: &Transaction<PrefixRecord, RepoDataRecord>,
+    ) -> Self {
+        match (
+            transaction.current_python_info.as_ref(),
+            transaction.python_info.as_ref(),
+        ) {
+            (Some(old), Some(new)) if old.short_version != new.short_version => {
+                PythonStatus::Changed {
+                    old: old.clone(),
+                    new: new.clone(),
+                }
+            }
+            (Some(_), Some(new)) => PythonStatus::Unchanged(new.clone()),
+            (None, Some(new)) => PythonStatus::Added { new: new.clone() },
+            (Some(old), None) => PythonStatus::Removed { old: old.clone() },
+            (None, None) => PythonStatus::DoesNotExist,
+        }
+    }
+
+    /// Returns the info of the current situation (e.g. after the transaction
+    /// completed).
+    pub(crate) fn current_info(&self) -> Option<&PythonInfo> {
+        match self {
+            PythonStatus::Changed { new, .. }
+            | PythonStatus::Unchanged(new)
+            | PythonStatus::Added { new } => Some(new),
+            PythonStatus::Removed { .. } | PythonStatus::DoesNotExist => None,
+        }
+    }
+
+    /// Returns the location of the python interpreter relative to the root of
+    /// the prefix.
+    pub(crate) fn location(&self) -> Option<&Path> {
+        Some(&self.current_info()?.path)
+    }
+}
