+++ src/lib.rs
@@ -2,7 +2,6 @@
 
 pub mod activation;
 pub mod cli;
-pub(crate) mod conda_prefix_updater;
 pub mod diff;
 pub mod environment;
 mod global;
@@ -20,6 +19,5 @@ mod build;
 mod rlimit;
 mod utils;
 
-pub use conda_prefix_updater::{CondaPrefixUpdated, CondaPrefixUpdater};
 pub use lock_file::{load_lock_file, UpdateLockFileOptions};
 pub use workspace::{DependencyType, Workspace, WorkspaceLocator};
