+++ src/lock_file/resolve/pypi.rs
@@ -50,6 +50,7 @@ use uv_resolver::{
 use uv_types::EmptyInstalledPackages;
 
 use crate::{
+    environment::CondaPrefixUpdated,
     lock_file::{
         records_by_name::HasNameVersion,
         resolve::{
@@ -63,7 +64,6 @@ use crate::{
     },
     uv_reporter::{UvReporter, UvReporterOptions},
     workspace::{Environment, EnvironmentVars},
-    CondaPrefixUpdated,
 };
 
 #[derive(Debug, thiserror::Error)]
